Updated nearest function
