+++ Model/Run.py
@@ -11,7 +11,6 @@ This is the Python file that you run to begin the simulation. Before you begin, 
 import Output
 import Simulation
 import Functions
-import Backend
 
 # setup() will create an instance of the Simulation class that holds all relevant information of the model.
 # this is done by reading a template .txt file that contains all initial parameters of the model.
@@ -19,7 +18,7 @@ simulation = Simulation.Simulation("C:\\Python37\\Seed Project\\Model\\template.
 
 # Will locate the diffusion points for the extracellular gradient to bins, used for chemotactic movement
 # called once as the locations don't change. not currently in use.
-Backend.setup_diffusion_bins(simulation)
+Functions.setup_diffusion_bins(simulation)
 
 # This will loop over all steps defined in the template file in addition to updating the current step
 # of the simulation.
@@ -42,7 +41,7 @@ for simulation.current_step in range(simulation.beginning_step, simulation.end_s
     Functions.update_queue(simulation)
 
     # Find the nearest NANOG high, GATA6 high, and differentiated cell within a fixed radius, used for movement
-    simulation.nearest()
+    Functions.nearest(simulation)
 
     # Locate the diffusion point (within a fixed radius) that has the highest FGF4 concentration.
     # not currently in use.
