+++ Model/template.txt
@@ -21,7 +21,7 @@ How many cells will be GATA6 high to begin the simulation? Ex. 247
 [ 0 ]
 
 How many cells will be NANOG high to begin the simulation? Ex. 91
-[ 10000 ]
+[ 1000 ]
 
 -----------------------------------------------------------------------------------------------------------------------
 Finite Dynamical System
@@ -58,7 +58,7 @@ running a simulation, type 1. Ex. 1
 [ 1 ]
 
 What will the final step number be? Ex. 30
-[ 5 ]
+[ 96 ]
 
 How long is each time step in seconds? Ex. 3600
 [ 1800 ]
@@ -187,7 +187,7 @@ Use FGF4 concentrations for NANOG high movements? Ex. False
 [ False ]
 
 Show the gradient color-map next to the image of the cell space? Ex. True
-[ False ]
+[ True ]
 
 The time has come. Execute Order Sixty-Six? Ex. True
 [ True ]
